% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_components.R
\name{default_merge_raw_cluster_args}
\alias{default_merge_raw_cluster_args}
\title{Default arguments for combining "raw clusters" into "aggregated raw clusters".}
\usage{
default_merge_raw_cluster_args()
}
\value{
A list with the following elements:

* \code{identical.cutoff}

* \code{nearly.identical.cutoff}

* \code{clustering.cutoff}

* \code{discard.singletons}
}
\description{
Default arguments for combining "raw clusters" into "aggregated raw clusters".
}
