% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ccc_from_hdp.R
\name{extract_ccc_from_hdp}
\alias{extract_ccc_from_hdp}
\title{Find the ccc and cdc that matched to a spectrum in ccc_0 and cdc_0
this function is to summarize the credint and mean of cccs and cdcs
and for further diagnostic plotting.}
\usage{
extract_ccc_from_hdp(signature, ccc_0, cos.merge = 0.9)
}
\arguments{
\item{signature}{The signature for which we want to get the intervals.}

\item{ccc_0}{A list XXXXXXXcontaining clust_categ_counts matrix from hdp.}

\item{cos.merge}{cosine similarity cutoff}
}
\description{
Find the ccc and cdc that matched to a spectrum in ccc_0 and cdc_0
this function is to summarize the credint and mean of cccs and cdcs
and for further diagnostic plotting.
}
