% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_components_from_clusters.R
\name{extract_components_from_clusters}
\alias{extract_components_from_clusters}
\title{Combine "raw clusters" of mutations in "components" (clusters of mutations generated by one mutational process).}
\usage{
extract_components_from_clusters(x, hc.cutoff = 0.1)
}
\arguments{
\item{x}{An \code{\link{hdpSampleChain-class}} or \
code{\link{hdpSampleMulti-class}} object}

\item{hc.cutoff}{the height to cut hierarchical clustering tree}
}
\value{
A list with the elements \describe{
\item{components}{Aggregated clusters as a data frame.
  Rows represent the categories (i.e. for
  mutational signature analysis, the mutation type,
  e.g. ACT -> AGT). Columns are
  aggregated clusters, i.e. clusters after all
  "raw clusters" across all Gibbs samples have been
  combined according to the divisive clustering.
  Each cell contains number of items (for mutational
  signature analysis, the number of mutations) of
  a particular category in a particular aggregated
  cluster.}

\item{components.post.samples}{A data frame with
  two columns: one is the index of column in
  \code{components} and
  the other is the number of posterior
  samples that contributed to that aggregated
  cluster (column in \code{components}).}

\item{components.cdc}{A numerical matrix.
   Each row is a Dirichlet process (DP).
   This can either be a leaf DP, which
   for mutational signatures corresponds
   to a biological sample (for exampl, a tumor),
   or an parent or ancestor DP.
   Each column corresponds to the cluster
   in the corresponding column n \code{components}}

\item{each.chain.noise.clusters}{Mo to document, called each.chain.noise.spectrum internally}

\item{each.chain.noise.cdc}{A matrix.
   Each row is a dirichlet process and
   each column corresponds to
   the cluster in \code{each.chain.noise.clusters}}

\item{multi.chains}{An \code{\link{hdpSampleChain-class}}
   or \code{\link{hdpSampleMulti-class}}
   object updated with component information.}

\item{nsamp}{The total number of posterior samples across all
  Gibbs sampling chains.}
}
}
\description{
Combine "raw clusters" of mutations in "components" (clusters of mutations generated by one mutational process).
}
\seealso{
\code{\link{hdp_posterior}}, \code{\link{hdp_multi_chain}},
 \code{\link{plot_comp_size}}, \code{\link{plot_comp_distn}},
 \code{\link{plot_dp_comp_exposure}}
}
