% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ccc_from_hdp.R
\name{extract_ccc_from_hdp}
\alias{extract_ccc_from_hdp}
\title{Find the ccc and cdc that matched to a spectrum in ccc_0 and cdc_0
this function is to summarize the credint and mean of cccs and cdcs
and for further diagnostic plotting.}
\usage{
extract_ccc_from_hdp(signature, ccc_0, cos.merge = 0.9)
}
\arguments{
\item{signature}{A numerical vector representing the signature for which
want to find information in \code{ccc_0}.}

\item{ccc_0}{"CCCs" from sample chains. A list of lists of numerical
matrices. At the top level, one list for each Gibbs sample chain.
Each element of a top-level list is a numerical matrix, with columns
for raw clusters and rows being mutation types. The number of
rows in these matrices should be the length of \code{signature}.}

\item{cos.merge}{The minimum cosine similarity for declaring a match
between \code{signature} and a column of one of the matrices in
in \code{ccc_0}. This should probably be the same as the cuttoff
for separating clusters of mutations after divisive clustering
\code{\link{extract_components}}.}
}
\value{
Invisibly, a list with at least the elements

- \code{ccc_mean}

- \code{ccc_credint}
}
\description{
Find the ccc and cdc that matched to a spectrum in ccc_0 and cdc_0
this function is to summarize the credint and mean of cccs and cdcs
and for further diagnostic plotting.
}
